refactor: remove Admin service
